fix(remnawave): align client with API v2.8.0 spec
- accept MONTH_ROLLING in trafficLimitStrategy
- CreateUserRequest.telegramId int64 (spec integer, avoids 32-bit truncation)
- Ping hits /api/system/health
- GetUsers stops on total
- generateUsername falls back to sha256[:16] when > 36 char pattern limit

+++ internal/remnawave/user_types.go
@@ -57,7 +57,7 @@ type CreateUserRequest struct {
 	ActiveInternalSquads []uuid.UUID `json:"activeInternalSquads,omitempty"`
 	ExternalSquadUuid    *uuid.UUID  `json:"externalSquadUuid,omitempty"`
 	Tag                  *string     `json:"tag,omitempty"`
-	TelegramID           *int        `json:"telegramId,omitempty"`
+	TelegramID           *int64      `json:"telegramId,omitempty"`
 	Description          *string     `json:"description,omitempty"`
 }
